A first person traversal game in Unreal Engine 5
This is a first person traversal puzzle game created in Unreal Engine 5 with the combination of C++ and Blueprint.

The level has three sections:
1) Dungeon
2) Maze
3) Crypt

The main goal of the game is to steal the golden statue from the crypt.

In the dungeon the player has to place a statue at a particular area to open a hidden door.

In the maze the player has to traverse and solve the maze to reach the crypt.

In the crypt the player has to steal the statue but work around the puzzle where a gate closes if the statue is moved.

Things learnt during the project:
1) Use of pointers and references in Unreal
2) Actor Components and Scene Components.
3) Lighting with Lumen
4) Modular level design
5) Line Tracing and Collision
6) Calling C++ functions from Blueprint.